Добавлена автоподстановка id текущего пользователя при переходе на страницу профиля
