+++ src/widget/Sidebar/UI/Sidebar/Sidebar.tsx
@@ -1,9 +1,10 @@
 import React, { memo, useMemo, useState } from 'react';
+import { useSelector } from 'react-redux';
 import { classNames } from 'shared/lib/classNames/classNames';
 import Button, { ButtonSize, ButtonTheme } from 'shared/UI/Button/Button';
 import LangSwitcher from 'shared/UI/LangSwitcher/LangSwitcher';
 import ThemeSwitcher from 'shared/UI/ThemeSwitcher/ThemeSwitcher';
-import { SidebarItemsList } from '../../model/SidebarItems';
+import { getSidebarItems } from 'widget/Sidebar/model/selectors/getSidebarItems';
 import SidebarItem from '../SidebarItem/SidebarItem';
 import cls from './Sidebar.module.scss';
 
@@ -13,8 +14,20 @@ interface SidebarProps {
 
 const Sidebar = memo(({ className }: SidebarProps) => {
   const [isCollapsed, setIsCollapsed] = useState(false);
+  const sidebarItemList = useSelector(getSidebarItems);
 
   const onToggle = () => setIsCollapsed((prev) => !prev);
+
+  const itemsList = useMemo(() => sidebarItemList.map((item) => (
+    <SidebarItem
+      key={item.text}
+      item={item}
+      collapsed={isCollapsed}
+    />
+  )), [
+    isCollapsed,
+    sidebarItemList,
+  ]);
 
   return (
     <div
@@ -33,13 +46,7 @@ const Sidebar = memo(({ className }: SidebarProps) => {
         {isCollapsed ? '>' : '<'}
       </Button>
       <div className={cls.items}>
-        {SidebarItemsList.map((item) => (
-          <SidebarItem
-            key={item.text}
-            item={item}
-            collapsed={isCollapsed}
-          />
-        ))}
+        {itemsList}
       </div>
       <div className={cls.switchers}>
         <ThemeSwitcher />
