--- src/widget/Sidebar/model/SidebarItems.ts
@@ -1,41 +1,0 @@
-import {
-  SVGProps,
-  VFC,
-} from 'react';
-import { RoutePath } from 'shared/config/routeConfig/routeConfig';
-import AboutIcon from 'shared/assets/icons/about.svg';
-import MainIcon from 'shared/assets/icons/home.svg';
-import ProfileIcon from 'shared/assets/icons/profile.svg';
-import ArticleIcon from 'shared/assets/icons/articles.svg';
-
-export interface ISidebarItem {
-	path: string
-	text: string
-	Icon: VFC<SVGProps<SVGSVGElement>>
-  authOnly?: boolean
-}
-
-export const SidebarItemsList: ISidebarItem[] = [
-  {
-    path: RoutePath.main,
-    text: 'Главная',
-    Icon: MainIcon,
-  },
-  {
-    path: RoutePath.about,
-    text: 'О сайте',
-    Icon: AboutIcon,
-  },
-  {
-    path: RoutePath.profile,
-    text: 'Профиль',
-    Icon: ProfileIcon,
-    authOnly: true,
-  },
-  {
-    path: RoutePath.articles,
-    text: 'Статьи',
-    Icon: ArticleIcon,
-    authOnly: true,
-  },
-];
