+++ src/widget/Sidebar/model/types/sidebar.ts
@@ -1,0 +1,8 @@
+import { SVGProps, VFC } from 'react';
+
+export interface ISidebarItem {
+	path: string
+	text: string
+	Icon: VFC<SVGProps<SVGSVGElement>>
+	authOnly?: boolean
+}
